refactor(extension): split io-launcher from new-active-feature-folder io.ts (#240)
F8 remediation: io.ts exceeded the 500-line limit (542). Extract the VS Code launcher seam
into io-launcher.ts (188 lines) and re-export from io.ts (now 386 lines). Behavior, public
exports, and message strings unchanged; no circular import. Full extension suite 85 suites /
999 tests pass; io.ts coverage 99.48% line / 91.04% branch, io-launcher.ts 97.87%/84.61%.
