+++ docs/features/active/2026-06-25-port-python-commands-to-typescript-240/evidence/qa-gates/remediation-final-toolchain.2026-06-26T05-25.md
@@ -1,0 +1,53 @@
+# Remediation Final Toolchain — F8 (Issue #240)
+
+Timestamp: 2026-06-26T05-25
+Working directory: extensions/drm-copilot/
+
+Loop executed in order: format -> lint -> type-check -> test (coverage). No stage changed files or failed; no restart required.
+
+## Format
+
+Command: `npx prettier --check "src/lib/new-active-feature-folder/**/*.ts" "test/lib/new-active-feature-folder/**/*.ts"`
+EXIT_CODE: 0
+Output Summary: All matched files use Prettier code style. Format clean.
+
+## Lint
+
+Command: `npm run lint` (eslint --no-error-on-unmatched-pattern src test)
+EXIT_CODE: 0
+Output Summary: 0 lint errors. No new suppressions added.
+
+## Type check
+
+Command: `npm run typecheck` (tsc -p ./ --noEmit)
+EXIT_CODE: 0
+Output Summary: 0 type errors.
+
+## Tests + Coverage
+
+Command: `node run-jest.cjs --coverage --collectCoverageFrom="src/lib/**/*.ts"`
+EXIT_CODE: 0
+Output Summary:
+- Test Suites: 85 passed, 85 total
+- Tests: 999 passed, 999 total
+- src/lib/** (All files): line 97.73%, branch 88.32% (baseline 97.73% / 88.29%; no regression — branch +0.03)
+- io.ts: line 99.48%, branch 91.04% (baseline 98.89% / 88.88%)
+- io-launcher.ts: line 97.87%, branch 84.61% (>= 85% line / >= 75% branch thresholds met)
+- index.ts: line 100% (re-export surface unchanged)
+- flow.ts: line 99.54%, branch 92.1% (unchanged)
+
+## File sizes
+
+Command: `wc -l src/lib/new-active-feature-folder/io.ts src/lib/new-active-feature-folder/io-launcher.ts`
+EXIT_CODE: 0
+Output Summary:
+- io.ts = 386 lines (< 500; was 542)
+- io-launcher.ts = 188 lines (< 500)
+
+## Result
+
+All gates green. Both files < 500 lines. Behavior preserved: launcher symbols
+(`INSIDERS_SIGNAL_NAMES`, `defaultEnvLookup`, `isInsidersSession`, `resolveCodeCli`,
+`CodeLauncherDeps`, `defaultCodeLauncher`, `defaultWhichLookup`) re-exported from `io.ts`
+via `./io-launcher`; `index.ts` and `flow.ts` imports unchanged; no regex, message string,
+or `gh --json` field-list change.
